ico2base64 moved to tapir.py
module that collects methods that are utilities for creating components
