+++ grasshopper-plugin/src/tapir.py
@@ -6,6 +6,7 @@ __all__ = ['Plugin', 'RhinoWrapper']
 # - - - - - - - - BUILT-IN IMPORTS
 import traceback, time
 import System
+import base64
 
 # - - - - - - - - LOCAL IMPORTS
 from tapir_py import parts, utility
@@ -101,3 +102,13 @@ class Factory(object):
             return True
         else:
             return False
+# ----- UTILLITIES
+def ico2base64(ico_file):
+    """
+    Converts icon image to base64 string 
+    to be used by get_internal_Icon_24x24 method
+    """
+    with open(ico_file, "rb") as image_file:
+        encoded_string = base64.b64encode(image_file.read())
+        image_file.close()
+    return encoded_string
